// frontend/src/api/models/SubtitleConfiguration.ts
export interface SubtitleConfiguration {
    /**
     * ID of the subtitle configuration.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    id?: string;
    /**
     * Kind of the subtitle configuration.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    kind?: string;
    /**
     * Name of the subtitle configuration.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    name?: string;
    /**
     * Name of the animation.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    animation?: string;
    /**
     * Color of the subtitles.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    color?: string;
    /**
     * Size of the subtitles.
     * @type {number}
     * @memberof SubtitleConfiguration
     */
    size?: number;
    /**
     * Font of the subtitles.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    font?: string;
    /**
     * Style of the subtitles, e.g., 'middle', 'bottom', 'top'.
     * @type {SubtitleStyle}
     * @memberof SubtitleConfiguration
     */
    subtitleStyle?: SubtitleStyle;
    /**
     * 
     * @type {SubtitleCapitalizationMethod}
     * @memberof SubtitleConfiguration
     */
    subtitleCapitalizationMethod?: SubtitleCapitalizationMethod;
    /**
     * 
     * @type {SubtitlePosition}
     * @memberof SubtitleConfiguration
     */
    subtitlePosition?: SubtitlePosition;
    /**
     * Background color for the subtitles.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    subtitleBoxFontColor?: string;
    /**
     * Transparency for the subtitles.
     * @type {number}
     * @memberof SubtitleConfiguration
     */
    subtitleBoxTransparency?: number;
    /**
     * 
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    subtitleBoxBackgroundColor?: string | null;
    /**
     * Color for the inactive part of the subtitles.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    subtitleInactiveColor?: string;
    /**
     * Color for the active (highlighted) part of the subtitles.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    subtitleActiveColor?: string;
    /**
     * Corner radius for the subtitles.
     * @type {number}
     * @memberof SubtitleConfiguration
     */
    subtitleBoxCornerRadius?: number;
    /**
     * Compensation factor for word width.
     * @type {number}
     * @memberof SubtitleConfiguration
     */
    subtitleBoxWidthCompensation?: number;
    /**
     * Border thickness for the subtitles.
     * @type {number}
     * @memberof SubtitleConfiguration
     */
    subtitleBoxBorderThickness?: number;
    /**
     * Font size for the subtitles.
     * @type {number}
     * @memberof SubtitleConfiguration
     */
    subtitleFontSize?: number;
    /**
     * Font name for the subtitles.
     * @type {string}
     * @memberof SubtitleConfiguration
     */
    subtitleFontName?: string;
    /**
     * Whether to use bold font for the subtitles.
     * @type {boolean}
     * @memberof SubtitleConfiguration
     */
    subtitleFontBold?: boolean;
    /**
     * Whether to use italic font for the subtitles.
     * @type {boolean}
     * @memberof SubtitleConfiguration
     */
    subtitleFontItalic?: boolean;
    /**
     * Whether to use underline font for the subtitles.
     * @type {boolean}
     * @memberof SubtitleConfiguration
     */
    subtitleFontUnderline?: boolean;
    /**
     * 
     * @type {LanguageCode}
     * @memberof SubtitleConfiguration
     */
    subtitleLanguageCode?: LanguageCode | null;
}